requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
langchain==0.0.225
flask
pypdf
python-dotenv
langchain_pinecone

-e .